backend/requirements.txt
fastapi
uvicorn
crewai
langchain
langchain-openai
openai
python-dotenv